Jumping bug fixed.  ***GameMaker Update***
***Update GameMaker before running, just in case!***

Fixed logic for destroying bomb objects by adding an alarm.  Previously,
bombs were set to be destroyed at the end of an animation, but only the
explosion is animated, not the bomb itself.  By changing the alarm time,
we can control how long the explosion "falls".  I added it to keep the
feel consistent with what we had before.  Anyway, this was the root of
the latest jumping problem.  Jumping works smoothly now, as long the the
physics world update speed isn't changed.